+++ app/controllers/application_controller.rb
@@ -1,14 +1,9 @@
 class ApplicationController < ActionController::API
   include ActionController::HttpAuthentication::Token::ControllerMethods
 
-  before_action :authenticate
-
-  def authenticate
-    authenticate_or_request_with_http_token do |token, _options|
-      User.find_by(token: token)
-  end
-
   def current_user
-    @current_user ||= User.find_by(token:  )
+    # メモ化（2回目以降、@current_userが未定義でnilなのか計算した結果がnilなのかを区別するため）
+    return @current_user if instance_variable_defined?(:@current_user)
+    @current_user = User.find_by(token: request.headers['Authorization'].split[1])
   end
 end
