+++ spec/requests/api/v1/books_request_spec.rb
@@ -1,0 +1,1 @@
+require 'rails_helper'
