+++ spec/requests/api/v1/users_request_spec.rb
@@ -1,16 +1,16 @@
 require 'rails_helper'
 
-RSpec.describe "Api::Vi::Users", type: :request do
+RSpec.describe "Api::V1::Users", type: :request do
   describe "GET /sing_up" do
     it "returns http success" do
-      get "/api/vi/users/sing_up"
+      get "/api/v1/users/sing_up"
       expect(response).to have_http_status(:success)
     end
   end
 
   describe "GET /login" do
     it "returns http success" do
-      get "/api/vi/users/login"
+      get "/api/v1/users/login"
       expect(response).to have_http_status(:success)
     end
   end
