--- spec/requests/books_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Books", type: :request do
-
-end
